回來發現有動過 不知道是啥 先上傳避免衝突

+++ MotaiProject/Models/PromotionRespoitory.cs
@@ -20,8 +20,8 @@ namespace MotaiProject.Models
             {
                 DetailPromotionViewModel Promo = new DetailPromotionViewModel();
                 Promo.pADimage = item.pADimage;
-                Promo.pCondition = item.pCondition;
-                Promo.pDiscount = item.pDiscount;
+                Promo.pCondition = (int)item.pCondition;
+                Promo.pDiscount = (double)item.pDiscount;
                 Promo.pPromotionDeadline = item.pPromotionDeadline;
                 Promo.pPromotionPostDate = item.pPromotionPostDate;
                 Promo.pPromotionStartDate = item.pPromotionStartDate;
@@ -41,8 +41,8 @@ namespace MotaiProject.Models
             tPromotion promotion = dbContext.tPromotions.FirstOrDefault(p => p.PromotionId == PromotionId);
             DetailPromotionViewModel Promo = new DetailPromotionViewModel();
             Promo.pADimage = promotion.pADimage;
-            Promo.pCondition = promotion.pCondition;
-            Promo.pDiscount = promotion.pDiscount;
+            Promo.pCondition = (int)promotion.pCondition;
+            Promo.pDiscount = (double)promotion.pDiscount;
             Promo.pPromotionDeadline = promotion.pPromotionDeadline;
             Promo.pPromotionPostDate = promotion.pPromotionPostDate;
             Promo.pPromotionStartDate = promotion.pPromotionStartDate;
